something's rotten in the kingdo of denmark
