+++ src/components/dataviz/Hemicycle/ParliamentChart.tsx
@@ -2,121 +2,15 @@
 
 import React, { useEffect, useRef } from 'react';
 import * as d3 from 'd3';
-
-// Interface for Party data
-export interface Party {
-  id: string;
-  name: string;
-  seats: number | any[];
-  color: string;
-}
+import { Party, Seat } from './types';
+import makeParliament from './parliament';
 
 // Interface for configuration
 interface Config {
   width: number;
   height: number;
-  mobileWidth: number;
-  mobileHeight: number;
   innerRadiusCoef: number;
 }
-
-// Interface for individual Seat data
-interface Seat {
-  polar: {
-    r: number;
-    teta: number;
-  };
-  cartesian: {
-    x: number;
-    y: number;
-  };
-  party?: Party;
-  data?: any;
-}
-
-// Helper function to calculate a series of values
-const series = (s: (i: number) => number, n: number): number => {
-  let r = 0;
-  for (let i = 0; i <= n; i++) {
-    r += s(i);
-  }
-  return r;
-};
-
-// Calculate the number of rows and other parameters for seats
-const calculateRows = (nSeats: number, innerRadiusCoef: number) => {
-  let nRows = 0;
-  let maxSeatNumber = 0;
-  let b = 0.5;
-
-  const calcFloor = (i: number) => Math.floor(Math.PI * (b + i));
-  const a = innerRadiusCoef / (1 - innerRadiusCoef);
-
-  while (maxSeatNumber < nSeats) {
-    nRows += 1;
-    b += a;
-    maxSeatNumber = series(calcFloor, nRows - 1);
-  }
-
-  return { nRows, maxSeatNumber, b };
-};
-
-// Create the seat objects with polar and cartesian coordinates
-const createSeats = (
-  nSeats: number,
-  nRows: number,
-  b: number,
-  rowWidth: number,
-  outerParliamentRadius: number,
-  innerParliamentRadius: number,
-) => {
-  const seats: Seat[] = [];
-  const seatsToRemove = nSeats - nSeats; // Placeholder logic; may need to adjust
-
-  for (let i = 0; i < nRows; i++) {
-    const rowRadius = innerParliamentRadius + rowWidth * (i + 0.5);
-    const rowSeats =
-      Math.floor(Math.PI * (b + i)) - Math.floor(seatsToRemove / nRows) - (seatsToRemove % nRows > i ? 1 : 0);
-    const anglePerSeat = Math.PI / rowSeats;
-
-    for (let j = 0; j < rowSeats; j++) {
-      const s: Seat = {
-        polar: {
-          r: rowRadius,
-          teta: -Math.PI + anglePerSeat * (j + 0.5),
-        },
-        cartesian: {
-          x: rowRadius * Math.cos(-Math.PI + anglePerSeat * (j + 0.5)),
-          y: rowRadius * Math.sin(-Math.PI + anglePerSeat * (j + 0.5)),
-        },
-      };
-      seats.push(s);
-    }
-  }
-
-  return seats;
-};
-
-// Assign the seats to the parties
-const assignSeatsToParties = (seats: Seat[], data: Party[]) => {
-  let partyIndex = 0;
-  let seatIndex = 0;
-
-  seats.forEach((s) => {
-    let party = data[partyIndex];
-    // const nSeatsInParty = typeof party.seats === 'number' ? party.seats : party.seats.length;
-
-    // if (seatIndex >= nSeatsInParty) {
-    //   partyIndex += 1;
-    //   seatIndex = 0;
-    //   party = data[partyIndex];
-    // }
-
-    s.party = party;
-    // s.data = typeof party.seats === 'number' ? null : party.seats[seatIndex];
-    seatIndex += 1;
-  });
-};
 
 interface ParliamentChartProps {
   data?: Party[];
@@ -130,12 +24,10 @@ const ParliamentChart: React.FC<ParliamentChartProps> = (props) => {
   const config: Config = props.config || {
     width: 600,
     height: 400,
-    mobileWidth: 300,
-    mobileHeight: 200,
     innerRadiusCoef: 0.2,
   };
 
-  const isMobile = window.innerWidth < 768;
+  // const isMobile = window.innerWidth < 768;
 
   const data: Party[] = props.data || [
     { id: 'leu', name: 'Free and Equal', seats: 59, color: '#700000' },
@@ -149,31 +41,10 @@ const ParliamentChart: React.FC<ParliamentChartProps> = (props) => {
     { id: 'brothers', name: 'Brothers of Italy', seats: 12, color: '#1d24bf' },
   ];
 
-  const makeParliament = (data: Party[], width: number, height: number, innerRadiusCoef: number) => {
-    const outerParliamentRadius = Math.min(width / 2, height);
-    const innerParliamentRadius = outerParliamentRadius * innerRadiusCoef;
-
-    let nSeats = data.reduce((acc, p) => acc + (typeof p.seats === 'number' ? Math.floor(p.seats) : p.seats.length), 0);
-
-    const { nRows, maxSeatNumber, b } = calculateRows(nSeats, innerRadiusCoef);
-
-    const rowWidth = (outerParliamentRadius - innerParliamentRadius) / nRows;
-    const seats = createSeats(nSeats, nRows, b, rowWidth, outerParliamentRadius, innerParliamentRadius);
-
-    seats.sort((a, b2) => a.polar.teta - b2.polar.teta || b2.polar.r - a.polar.r);
-
-    assignSeatsToParties(seats, data);
-
-    return { seats, rowWidth };
-  };
+  console.log('Data:', data);
 
   useEffect(() => {
-    const { seats, rowWidth } = makeParliament(
-      data,
-      isMobile ? config.mobileWidth : config.width,
-      isMobile ? config.mobileHeight : config.height,
-      config.innerRadiusCoef,
-    );
+    const { seats, rowWidth } = makeParliament({ data, ...config });
 
     const seatRadius = (d: Seat) => {
       let r = 0.4 * rowWidth;
@@ -185,18 +56,11 @@ const ParliamentChart: React.FC<ParliamentChartProps> = (props) => {
 
     const svg = d3
       .select(svgRef.current)
-      .attr('width', isMobile ? config.mobileWidth : config.width)
-      .attr('height', isMobile ? config.mobileHeight : config.height)
+      // .attr('width', isMobile ? config.mobileWidth : config.width)
+      // .attr('height', isMobile ? config.mobileHeight : config.height)
       .style('background-color', '#f8f7f3');
 
-    const group = svg
-      .append('g')
-      .attr(
-        'transform',
-        `translate(${isMobile ? config.mobileWidth / 2 : config.width / 2}, ${
-          isMobile ? config.mobileHeight / 2 + 80 : config.height / 2 + 150
-        })`,
-      );
+    const group = svg.append('g').attr('transform', `translate(${config.width / 2}, ${config.height / 2 + 150})`);
 
     group
       .selectAll('.seat')
@@ -208,7 +72,7 @@ const ParliamentChart: React.FC<ParliamentChartProps> = (props) => {
       .attr('cy', (d) => d.cartesian.y)
       .attr('fill', (d) => d.party?.color || 'gray')
       .attr('r', seatRadius);
-  }, [data, isMobile, config]);
+  }, [data, config]);
 
   return <svg ref={svgRef}></svg>;
 };
