+++ src/components/dataviz/Hemicycle/types.ts
@@ -1,6 +1,19 @@
 export type Party = {
   id: string;
   name: string;
-  seats: number;
+  seats: number | any[];
   color: string;
+};
+
+export type Seat = {
+  polar: {
+    r: number;
+    teta: number;
+  };
+  cartesian: {
+    x: number;
+    y: number;
+  };
+  party?: Party;
+  data?: any;
 };
